+++ packages/site/src/componentList.ts
@@ -1,0 +1,104 @@
+export const componentList = [
+  {
+    title: "AnimatedPresence",
+    to: "/components/AnimatedPresence",
+    imageURL: "/Animation.png",
+    additionalMatches: ["Motion", "Animation", "Reveal"],
+  },
+  {
+    title: "AnimatedSwitcher",
+    to: "/components/AnimatedSwitcher",
+    imageURL: "/Animation.png",
+    additionalMatches: ["Motion", "Animation"],
+  },
+  {
+    title: "Autocomplete",
+    to: "/components/Autocomplete",
+    imageURL: "/Autocomplete.png",
+  },
+  {
+    title: "Banner",
+    to: "/components/Banner",
+    imageURL: "/Banner.png",
+  },
+  {
+    title: "Button",
+    to: "/components/Button",
+    imageURL: "/Button.png",
+  },
+  {
+    title: "ButtonDismiss",
+    to: "/components/ButtonDismiss",
+    imageURL: "/ButtonDismiss.png",
+    additionalMatches: ["Close", "Dismiss", "Remove"],
+  },
+  {
+    title: "Checkbox",
+    to: "/components/Checkbox",
+    imageURL: "/Checkbox.png",
+  },
+  {
+    title: "Chip",
+    to: "/components/Chip",
+    imageURL: "/Chip.png",
+    additionalMatches: ["Pill", "Badge", "Tag"],
+  },
+  {
+    title: "Chips",
+    to: "/components/Chips",
+    imageURL: "/Chips.png",
+    additionalMatches: ["Pill", "Badge", "Tag"],
+  },
+  {
+    title: "Countdown",
+    to: "/components/Countdown",
+    imageURL: "/Countdown.png",
+  },
+  {
+    title: "Disclosure",
+    to: "/components/Disclosure",
+    imageURL: "/Disclosure.png",
+    additionalMatches: [
+      "Accordion",
+      "Collapsable",
+      "Expandable",
+      "Toggle",
+      "Reveal",
+    ],
+  },
+  {
+    title: "Emphasis",
+    to: "/components/Emphasis",
+    imageURL: "/Emphasis.png",
+    additionalMatches: ["Highlight", "Strong", "Bold", "Italic"],
+  },
+  {
+    title: "Heading",
+    to: "/components/Heading",
+    imageURL: "/Heading.png",
+    additionalMatches: ["Title", "Subtitle", "Header"],
+  },
+  {
+    title: "Icon",
+    to: "/components/Icon",
+    imageURL: "/Icon.png",
+  },
+  {
+    title: "Link",
+    to: "/components/Link",
+    imageURL: "/Link.png",
+    additionalMatches: ["Anchor", "URL"],
+  },
+  {
+    title: "StatusLabel",
+    to: "/components/StatusLabel",
+    imageURL: "/StatusLabel.png",
+    additionalMatches: ["Badge", "Pill", "Tag", "InlineLabel", "Inline Label"],
+  },
+  {
+    title: "Switch",
+    to: "/components/Switch",
+    imageURL: "/Switch.png",
+    additionalMatches: ["Toggle"],
+  },
+];